Normalize layout handling from markdown files (#5)
This update allows specifying the name of a layout without specifying its path or file extension. The `normalizeLayouts` remark plugin will update the layout import path before its processed by astro. The tsconfig is updated with shortcuts for layouts (and components) so the actual path doesn't have to be referenced.