Fix build, drop content
Had to update my-remark to restore the `@layout/` prefixing for layout strings because the path was getting broken for committed files (like the 404 page). 

The integration for astro-md just isn't going to work. I'm going just declare bankrupcy on it for now. Will simplify it down in the next commit.