Get site building (#11)
* WIP: Node Compat

As I was prepping for deployment to cloudflare pages I realized that the build was failing due to cloudflare not supporting certain node builtins that were being used by some dependencies.

* Refactor to workspace; custom remark links

* Finish implement obsidian wiki-links

* Fix build, drop content

Had to update my-remark to restore the `@layout/` prefixing for layout strings because the path was getting broken for committed files (like the 404 page). 

The integration for astro-md just isn't going to work. I'm going just declare bankrupcy on it for now. Will simplify it down in the next commit.

* Minor packaging tweaks