Ensure node:fs isn't pulled in
The `staticFileResponse` function was being included in the build even though it's not used in a prod build which lead to astro trying to include node:fs in the bundle which resulted in a build error. I've instead wrapped the implementation of the `staticFileResponse` function in a if statement checking for NODE_ENV = development which should ensure the contents aren't included in the final build.